+++ src/include/S3/S3ListPartsRequest.h
@@ -19,12 +19,11 @@
 
 /** Contains the parameters used for the listParts operation.
  *
- * \ingroup S3
  */
 @interface S3ListPartsRequest:S3Request {
-    NSString *uploadId;
-    int      maxParts;
-    int      partNumberMarker;
+    NSString  *uploadId;
+    NSInteger maxParts;
+    NSInteger partNumberMarker;
 }
 
 /** The uploadId of the multipart upload to list */
