+++ src/include/S3/S3Part.h
@@ -17,12 +17,11 @@
 
 /** a container for elements related to a particular part of a multipart upload.
  *
- * \ingroup s3
  */
 @interface S3Part:NSObject {
     NSString  *etag;
     NSDate    *lastModified;
-    int       partNumber;
+    NSInteger partNumber;
     long long size;
 }
 
@@ -33,7 +32,7 @@
 @property (nonatomic, retain) NSDate *lastModified;
 
 /** The unique part number identifying the part. */
-@property (nonatomic) int partNumber;
+@property (nonatomic) NSInteger partNumber;
 
 /** The size of the uploaded part data. */
 @property (nonatomic) long long size;
